backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
blinker==1.9.0
celery==5.5.3
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
num2words==0.5.14
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
platformdirs==4.3.7
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
reportlab==4.4.4
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.14
Werkzeug==3.1.3
